requirements.txt
google-genai>=1.0.0
feedparser>=6.0.11
requests>=2.32.0
python-dotenv>=1.0.1
PyYAML>=6.0.2
